manual_reviewer: live-test fix-pass round 2 + UI/UX hotkey fixes
Round-2 critical fixes from multi-agent code review (351/351 tests green)
plus live-test bugs surfaced during 2026-04-29 session against running
LS + sam3_1 + ml_backend stack.

Key fixes:
- ml_backend: V-tool RectangleLabels smartOnly=true, original_width/height
  on smart routes, retry-once transport wrapper, 4xx redaction, label snap
  to palette, smart_click prefers triggering keypoint label
- pipeline_io: WAL pragmas on read+write paths, bbox clamp+warn,
  path-traversal None+log, _safe_int on null LS ids
- scripts: build_tasks chunked POST + 5xx exponential backoff +
  strict_existing fail-loud; export_to_aa_v4 per-step try/except,
  composite-key trace dedup fallback for null ls_completion_id,
  --in-file/--since mutex; run_reconcile per-group flush + circuit
  breaker + KeyboardInterrupt drain; mark_dedup required mutex group
- reconcile: anchor-greedy clustering w/ rounded confidence,
  seed_iou against best member, structured reject_reason enum
- aav4: viewer registers human_review + reconcile in PIPELINE_STAGES
  + _BUILDERS; ReconciledDetection.reject_reason added; sam3_1
  --gpu cuda:N parsing fixed; visual_prompt BF16 -> float32 before .numpy()
- XML: pinned 24 hotkeys across all 3 label palettes (1-0,q-w-e-r-t-y-
  u-i-o-p,a-s-d-f) avoiding "v" so V-tool hotkey wins
- ml_backend server: _summarize_context now logs from_name:type so V-tool
  drafts can be distinguished from regular bbox draws

New tests: test_boundary.py (boundary round-trip + concurrent W+R),
test_build_tasks.py (chunking + retry + redaction + idempotency)

Live-test docs: docs/live_test_plan.md (12-workflow runbook),
docs/codebase_review_2026-04-29_live.md (results scoreboard + 2 live
bug fixes + 1 viewer schema gap + 4 ops findings)

+++ data_miner/auto_annotation_v4/models/sam3_1.py
@@ -463,12 +463,14 @@ class SAM3OneModel(BaseDetectorModel):
 
         if boxes_t is None or scores_t is None:
             return SAM3VisualPromptResponse(boxes_norm=[], scores=[])
+        # SAM 3.1 returns BFloat16 tensors on GPU; cast to float32 before
+        # .numpy() since NumPy has no BF16 dtype.
         if isinstance(boxes_t, torch.Tensor):
-            boxes_px = boxes_t.detach().cpu().numpy()
+            boxes_px = boxes_t.detach().to(torch.float32).cpu().numpy()
         else:
             boxes_px = np.asarray(boxes_t)
         if isinstance(scores_t, torch.Tensor):
-            scores_np = scores_t.detach().cpu().numpy()
+            scores_np = scores_t.detach().to(torch.float32).cpu().numpy()
         else:
             scores_np = np.asarray(scores_t)
         if boxes_px.size == 0:
